pyproject.toml
[project]
name = "no-emoji"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "mysql-connector-python>=9.4.0",
    "pyjwt[crypto]>=2.10.1",
    "redis>=6.4.0",
    "websockets>=15.0.1",
]
